tests: move butane dl to confftest.py so that it is available for all tests
